유니티 클라이언트 수정

+++ Client/Assets/Network/Datas/Protocols.cs
@@ -1,7 +1,7 @@
 using System;
 
 [Serializable]
-public enum Protocols : short
+public enum Protocols : byte
 {
 	NONE = 0,
